chore: Update user routes and JWT generation

+++ internal/utils/jwt_util.go
@@ -13,11 +13,11 @@ import (
 var jwtKey = []byte("your-secret-key")
 
 type Claims struct {
-	UserID int64 `json:"user_id"`
+	UserID int `json:"user_id"`
 	jwt.StandardClaims
 }
 
-func GenerateJWT(userID int64) (string, error) {
+func GenerateJWT(userID int) (string, error) {
 	claims := &Claims{
 		UserID: userID,
 		StandardClaims: jwt.StandardClaims{
